[FLINK-1159] Refactor packages layout
