--- flink-scala/src/test/scala/org/apache/flink/api/scala/extensions/acceptPartialFunctions/data/KeyValuePair.scala
@@ -1,9 +1,0 @@
-package org.apache.flink.api.scala.extensions.impl.acceptPartialFunctions.data
-
-/**
-  * Simple case class to test the `acceptPartialFunctions` extension
- *
-  * @param id A numerical identifier
-  * @param value A textual value
-  */
-private[acceptPartialFunctions] case class KeyValuePair(id: Int, value: String)
